+++ scripts/build-unix.sh
@@ -218,5 +218,7 @@ if [ "$target" = test ]; then
 	"$root/scripts/test-encoding.sh" "$src/jvim3" || rc=1
 	echo
 	"$root/scripts/test-editing.sh" "$src/jvim3" || rc=1
+	echo
+	"$root/scripts/test-syntax.sh" "$src/jvim3" || rc=1
 	exit $rc
 fi